Add support for nested menus
This allows entry commands to be placed inside another entry in order
to build menu trees. This can work in two ways.

Firstly, entries inside an entry which does not have an OS loader will
result in the parent entry being a link which, when selected, will open
another menu containing the entries inside it.

Secondly, entries inside an entry which *does* have an OS loader will
be presented as additional functions that can be accessed through the
F* keys, and the parent entry itself will still boot the OS when Enter
is pressed on it. An example use for this is that many Linux distros
keep older kernel versions around when a new version is installed, and
typically fill up the GRUB or whatever menu with these. Using nested
menus, it would be possible to relegate these to a separate menu, and
keep only one entry for the distro in the main boot menu.

Example usage:

    entry "Arch Linux" {
        device "uuid:..."

        entry "Options" {
            entry "Fallback" {
                linux \
                    "/boot/vmlinuz-linux rw root=UUID=${device_uuid}" \
                    ["/boot/initramfs-linux-fallback.img"]
            }
        }

        linux \
            "/boot/vmlinuz-linux rw root=UUID=${device_uuid}" \
            ["/boot/initramfs-linux.img"]
    }

This results in a "Arch Linux" entry in the main menu that has an
"Options" menu accessible by pressing F2 when the entry is selected.
The "Options" menu has a single "Fallback" entry which boots using the
fallback initramfs.

In this commit some modifications have been made to other standard keys.
F10 now brings up the shell rather than F2, and F9 brings up the debug
log rather than F10.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>