EFI: CD booting support
No major changes needed. Just changed boot directory detection so that
it doesn't get done for CDs (or network devices). This is because
creating a bootable CD for EFI requires embedding a FAT filesystem
image to act as an EFI system partition as an El Torito boot image.
The file path we receive from the firmware will refer to the boot path
within the image, but we do not use that image for anything in KBoot,
we instead mount the real ISO9660 filesystem on the CD.

Also added a QEMU test script for CD booting.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>